fixed the discover block design
